refactor: get the cdc lib to build.
