+++ integration-libs/cdc/cdc.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core';
-import { CdcComponentsModule } from '@spartacus/cdc/components';
 import { CdcCoreModule } from '@spartacus/cdc/core';
 
 @NgModule({
   declarations: [],
-  imports: [CdcCoreModule, CdcComponentsModule],
+  imports: [CdcCoreModule],
 })
 export class CdcModule {}
