--- integration-libs/cdc/components/public_api.ts
@@ -1,1 +1,0 @@
-export * from './index';
