+++ integration-libs/cdc/core/cdc-core.module.ts
@@ -1,9 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { CdcAuthModule } from './auth/cdc-auth.module';
 
 @NgModule({
   declarations: [],
-  imports: [CommonModule, CdcAuthModule],
+  imports: [CommonModule],
 })
 export class CdcCoreModule {}
