+++ integration-libs/cdc/root/cdc-root.module.ts
@@ -2,6 +2,8 @@ import { CommonModule } from '@angular/common';
 import { APP_INITIALIZER, ModuleWithProviders, NgModule } from '@angular/core';
 import { ConfigInitializerService, provideConfig } from '@spartacus/core';
 import { LogoutGuard } from '@spartacus/storefront';
+import { CdcAuthModule } from './auth/cdc-auth.module';
+import { CdcComponentsModule } from './components/cdc-components.module';
 import { CdcConfig } from './config/cdc-config';
 import { CdcJsService } from './facade/cdc-js.service';
 import { CdcLogoutGuard } from './guards/cdc-logout.guard';
@@ -20,7 +22,7 @@ export function cdcJsFactory(
 @NgModule({
   declarations: [],
   providers: [{ provide: LogoutGuard, useExisting: CdcLogoutGuard }],
-  imports: [CommonModule],
+  imports: [CommonModule, CdcAuthModule, CdcComponentsModule],
 })
 export class CdcRootModule {
   static forRoot(config: CdcConfig): ModuleWithProviders<CdcRootModule> {
