+++ integration-libs/cdc/root/components/gigya-raas/gigya-raas.component.scss
