+++ integration-libs/cdc/root/components/index.ts
@@ -1,2 +1,3 @@
 export * from './cdc-components.module';
 export * from './gigya-raas/index';
+export * from './models/cms.model';
