+++ integration-libs/cdc/root/index.ts
@@ -1,3 +1,5 @@
+export * from './auth/facade/index';
+export * from './auth/store/actions/index';
 export * from './cdc-root.module';
 export * from './config/index';
 export * from './facade/index';
