+++ projects/storefrontapp/tsconfig.server.json
@@ -135,9 +135,6 @@
       ],
       "@spartacus/tracking/tms": ["../../feature-libs/tracking/tms/public_api"],
       "@spartacus/cdc": ["../../integration-libs/cdc/public_api"],
-      "@spartacus/cdc/components": [
-        "../../integration-libs/cdc/components/public_api"
-      ],
       "@spartacus/cdc/core": ["../../integration-libs/cdc/core/public_api"],
       "@spartacus/cdc/root": ["../../integration-libs/cdc/root/public_api"],
       "@spartacus/cds": ["../../integration-libs/cds/public_api"],
